requirements.txt
azure-ai-projects>=2.0.0a20250915020
openai
azure-identity
python-dotenv